pyproject.toml
#################################################
# Project metadata
#################################################

[project]
requires-python = ">=3.11"
name = "ip_mensageria_alocacao_api"
version = "0.1.0"
description = "A minimal FastAPI project template."
readme = "README.md"
keywords = ["Comunicação em Saúde", "Saúde Digital", "Sistemas de Alerta"]
license-files = ["LICENSE"]
authors = [{name = "ImpulsoGov", email = "contato@impulsogov.com"}]
dependencies = [
    "catboost>=1.2.0",
    "bcrypt>=4.2.0",
    "fastapi>=0.115.4",
    "gunicorn>=23.0.0",
    "pandas>=2.0.3",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.17",
    "uvicorn>=0.32.0",
    "google-cloud-bigquery>=3.39.0",
    "google-cloud-storage>=2.18.0",
    "pytest-asyncio>=1.3.0",
    "libpass[bcrypt]>=1.9.3",
    "scikit-learn==1.6.0",
]
[dependency-groups]
dev = [
    "httpx>=0.27.2",
    "mypy>=1.13.0",
    "pip-tools>=7.4.1",
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "requests>=2.32.3",
    "ruff>=0.7.1",
]

[project.urls]
Repository = "https://github.com/ImpulsoGov/ip-mensageria-alocacao-api.git"

#################################################
# Mypy config
#################################################

[tool.mypy]
follow_imports = "skip"
ignore_missing_imports = true
warn_no_return = true
warn_unused_ignores = true
allow_untyped_globals = true
allow_redefinition = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = "ip_mensageria_alocacao_api.tests.*"
ignore_errors = true

#################################################
# Ruff config
#################################################

[tool.ruff]
respect-gitignore = true
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes `E` and `F` codes by default
select = ["E", "F", "PT", "C4", "I"]
ignore = ["E501"]

per-file-ignores = {}

[tool.ruff.lint.mccabe]
max-complexity = 10

#################################################
# Pytest config
#################################################

[tool.pytest.ini_options]
addopts = "--strict-markers --maxfail 1 --cov src tests/ --no-header"
markers = """
    integration: mark a test as an integration test.
    asyncio: mark a test as an asyncio test.
"""
console_output_style = "progress"

#################################################
# Setuptools config
#################################################

[tool.setuptools.packages.find]
where = ["src", "."]
